Added automation recording and screenshots
